feat(nx-micronaut): add dynamic prompt to fetch micronaut features list
BREAKING CHANGE: Nx `v15.8.x` is now the minimum required version to run the plugin

We now leverage Nx's new `NX_INTERACTIVE` environment variable to check whether we are running in interactive mode (normal cli) or not.
When true, we automatically fetch `Micronaut` features and present them in an **autocomplete** prompt with **multi-select** support,
so you can easily select which ones you want to include in your project.